backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=11.0
opencv-python>=4.8.0
numpy>=1.21.0
motor>=3.0.0
pymongo>=4.0.0
groq>=0.4.0
requests>=2.28.0
python-multipart>=0.0.5
Pillow>=9.0.0
python-dotenv>=0.19.0
qrcode[pil]>=7.4.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.25.0
safetensors>=0.4.0